02-semantic-tools-demo/requirements.txt
strands-agents[openai]>=1.27.0
sentence-transformers>=3.4.0
faiss-cpu>=1.9.0
neo4j>=5.28.0
python-dotenv>=1.0.1
